--- src/main/java/com/example/foodduck/exception/ExpiredShoppingCartException.java
@@ -1,7 +1,0 @@
-package com.example.foodduck.exception;
-
-public class ExpiredShoppingCartException extends RuntimeException {
-    public ExpiredShoppingCartException(String message) {
-        super(message);
-    }
-}
