--- src/main/java/com/example/foodduck/exception/StoreMismatchException.java
@@ -1,7 +1,0 @@
-package com.example.foodduck.exception;
-
-public class StoreMismatchException extends RuntimeException {
-    public StoreMismatchException(String message) {
-        super(message);
-    }
-}
