+++ src/main/java/com/example/foodduck/exception/custom/ExpiredShoppingCartException.java
@@ -1,0 +1,9 @@
+package com.example.foodduck.exception.custom;
+
+import org.springframework.http.HttpStatus;
+
+public class ExpiredShoppingCartException extends ApplicationException {
+    public ExpiredShoppingCartException(String message) {
+        super(message, HttpStatus.BAD_REQUEST);
+    }
+}
