+++ src/main/java/com/example/foodduck/exception/custom/StoreMismatchException.java
@@ -1,0 +1,9 @@
+package com.example.foodduck.exception.custom;
+
+import org.springframework.http.HttpStatus;
+
+public class StoreMismatchException extends ApplicationException {
+    public StoreMismatchException(String message) {
+        super(message, HttpStatus.BAD_REQUEST);
+    }
+}
